move to consensus.rs
